Use native LSP features for diagnostics
Also load language servers correctly. Apparently we're not supposed to
do so from filetype plugins, even though that makes a lot of sense to my
mind.

This has sort of made ALE redundant, but I'm unsure if OmniSharp works
properly without it, so C# development might be a pain.
I've disabled ALE for now, and will need to look into it.