Don't start element-desktop on startup
Electron seems to first display a loading window, which is then
destroyed, after which the actual chat window is displayed.
This means that, even though I tell qtile to start the program in the
"cht" group, the actual chat window is rendered on whatever my default
group is.

Fucking electron.